Added ability to use suffixes in match strings
It's necessary to use suffix sometimes when there are unintended partial matches due to missing json keys in the pattern json object. 
For example, be able to replace matches like @@env.css.darkStyle@@ instead of just @@env.css.darkStyle. This is needed to avoid erroneously substituting a partial match like @@env.css.darkStyle_header with the @@env.css.darkStyle match when darkStyle_header is not found in the pattern json.
5793b26